Add binary for generating admin stats from DB.
slangbrain-stat is run with every backup now.
Added migrations fixing the current data
and added new buckets to track imports and notfies.

+++ integration/profile_cache_test.go
@@ -14,9 +14,9 @@ import (
 
 type profile struct{}
 
-func (p profile) Name() string      { return "Martin" }
-func (p profile) Locale() string    { return "en_US" }
-func (p profile) Timezone() float64 { return 2 }
+func (p profile) Name() string   { return "Martin" }
+func (p profile) Locale() string { return "en_US" }
+func (p profile) Timezone() int  { return 2 }
 
 func TestProfileCache(t *testing.T) {
 	store, cleanup := initDB(t)
